remove non-util utilities
