+++ martin/src/config/file/pg/mod.rs
@@ -6,3 +6,5 @@ pub use config_function::*;
 
 mod config_table;
 pub use config_table::*;
+
+mod utils;
